pyproject.toml
[project]
name = "yt-llm-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "openai-whisper>=20250625",
    "python-dotenv>=1.2.1",
    "pytubefix>=10.3.5",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "youtube-transcript-api>=1.2.3",
]
